// worktimetracker.client/src/generate-api/models/CreateDeviceCommand.ts
/**
 * WorkTimeTracker.Server
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { DeviceStatus } from '../models/DeviceStatus';
import { HttpFile } from '../http/http';

export class CreateDeviceCommand {
    'name': string;
    'description'?: string | null;
    'location'?: string | null;
    'status'?: DeviceStatus;
    'assignedUserId'?: string | null;
    'deviceCategoryIds'?: Array<number> | null;

    static readonly discriminator: string | undefined = undefined;

    static readonly mapping: {[index: string]: string} | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "name",
            "baseName": "name",
            "type": "string",
            "format": ""
        },
        {
            "name": "description",
            "baseName": "description",
            "type": "string",
            "format": ""
        },
        {
            "name": "location",
            "baseName": "location",
            "type": "string",
            "format": ""
        },
        {
            "name": "status",
            "baseName": "status",
            "type": "DeviceStatus",
            "format": ""
        },
        {
            "name": "assignedUserId",
            "baseName": "assignedUserId",
            "type": "string",
            "format": "uuid"
        },
        {
            "name": "deviceCategoryIds",
            "baseName": "deviceCategoryIds",
            "type": "Array<number>",
            "format": "int32"
        }    ];

    static getAttributeTypeMap() {
        return CreateDeviceCommand.attributeTypeMap;
    }

    public constructor() {
    }
}